Refactor(consensus): acceptor and votingmachine are in separate packages.
