--- protocol/consensus/proposer.go
@@ -1,8 +1,0 @@
-package consensus
-
-type Proposer struct {
-}
-
-func NewProposer() *Proposer {
-	return &Proposer{}
-}
